dynamic  and middle ware

+++ routes.php
@@ -1,15 +1,16 @@
 <?php
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listing/edit/{id}', 'ListingController@edit');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listing/edit/{id}', 'ListingController@edit', ['auth']);
 $router->get('/listing/{id}', 'ListingController@show');
 
 $router->post('/listings', 'ListingController@store');
-$router->put('/listing/{id}', 'ListingController@update');
-$router->delete('/listing/{id}', 'ListingController@destroy');
+$router->put('/listing/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listing/{id}', 'ListingController@destroy', ['auth']);
 
-$router->get('/auth/register', 'UserController@register');
-$router->post('/auth/register', 'UserController@store');
-$router->get('/auth/login', 'UserController@login');
-$router->post('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@register', ['guest']);
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
